- try to find the source of this blocking bug

+++ ROSCode/sketchbook/robotica/src/sketch.ino
@@ -76,8 +76,8 @@ void move(double distance,int speed);
 // ROS Related
 std_msgs::Float32 sonar_msg;
 std_msgs::Float32 motorVolt_msg;
-std_msgs::Int32   encoder1_msg;
-std_msgs::Int32   encoder2_msg;
+std_msgs::Float32 encoder1_msg;
+std_msgs::Float32 encoder2_msg;
 geometry_msgs::TransformStamped robotPoseMsg;
 
 ros::Publisher pub_sonar    ("sonar"   ,  &sonar_msg);
@@ -372,7 +372,7 @@ void loop()
     encoder1_msg.data  = lastEncoder1Val;
     pub_encoder1.publish(&encoder1_msg);
     encoder2_msg.data  = lastEncoder2Val;
-    pub_encoder2.publish(&encoder2_msg); 
+    pub_encoder2.publish(&encoder2_msg);      
 
     sonar_msg.data     = sonarDist();
     nh.spinOnce();
